+++ CHANGES.rst
@@ -1,6 +1,11 @@
+0.3
+---
+- Add queue.Pusher, queue.Puller, queue.Publisher, queue.Subscriber
+- Add repository.get_github_ssh_url
+
 0.2
 ---
-- Add log.format_relative_path
+- Add log.format_path
 - Add text.remove_punctuation
 
 0.1